slider i meny och paus, ändrat font mm
Eftersom jag inte har fått till att musiken ska fortsätta mellan
scenerna och det ända som fanns i options var att ändra volymen så har
jag flyttat slidern till startmenyn, har även lagt till en slider i
pausmenyn så man kan ändra volymen där. Testade att lägga in en "pixel
font" och ändrade all text till den. Poängen visas nu i heltal, ett
poäng per sekund.

+++ Assets/scripts/GameplayScripts/HudScript.cs
@@ -7,9 +7,7 @@ public class HudScript : MonoBehaviour {
 	PausMenu pause;
 
 	public static float playerScore;
-
     Text text;
-
 
     public void Start()
     {
@@ -19,7 +17,7 @@ public class HudScript : MonoBehaviour {
 	public void Update () 
 	{   
         playerScore += Time.deltaTime;
-        text.text = "" + (playerScore );
+        text.text = "" + ((int)playerScore );
 	}
 
 	public static void IncreaseScore(float amount)
